refactor: AI 생성 문제, 사용자 생성 문제의 상위 패키지 생성 및 패키지 이동
