+++ src/main/java/com/app/domain/categorizedproblem/controller/CategorizedProblemController.java
@@ -4,7 +4,7 @@ import com.app.domain.categorizedproblem.dto.CategorizedProblemDto;
 import com.app.domain.categorizedproblem.entity.CategorizedProblem;
 import com.app.domain.categorizedproblem.mapper.CategorizedProblemMapper;
 import com.app.domain.categorizedproblem.service.CategorizedProblemService;
-import com.app.domain.membersavedproblem.dto.MemberSavedProblemDto;
+import com.app.domain.problem.membersavedproblem.dto.MemberSavedProblemDto;
 import com.app.domain.membersavedsummary.dto.MemberSavedSummaryDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
