+++ src/main/java/com/app/domain/categorizedproblem/service/CategorizedProblemService.java
@@ -6,13 +6,13 @@ import com.app.domain.category.entity.Category;
 import com.app.domain.category.service.CategoryService;
 import com.app.domain.member.entity.Member;
 import com.app.domain.member.service.MemberService;
-import com.app.domain.membersavedproblem.dto.MemberSavedProblemDto;
-import com.app.domain.membersavedproblem.entity.MemberSavedProblem;
-import com.app.domain.membersavedproblem.mapper.MemberSavedProblemMapper;
-import com.app.domain.membersavedproblem.service.MemberSavedProblemService;
+import com.app.domain.problem.membersavedproblem.dto.MemberSavedProblemDto;
+import com.app.domain.problem.membersavedproblem.entity.MemberSavedProblem;
+import com.app.domain.problem.membersavedproblem.mapper.MemberSavedProblemMapper;
+import com.app.domain.problem.membersavedproblem.service.MemberSavedProblemService;
 import com.app.domain.membersavedsummary.dto.MemberSavedSummaryDto;
-import com.app.domain.problem.entity.AiGeneratedProblem;
-import com.app.domain.problem.service.ProblemService;
+import com.app.domain.problem.aigeneratedproblem.entity.AiGeneratedProblem;
+import com.app.domain.problem.aigeneratedproblem.service.ProblemService;
 import com.app.global.config.ENUM.ProblemType;
 import com.app.global.error.ErrorCode;
 import com.app.global.error.exception.BusinessException;
