+++ src/main/java/com/app/domain/problem/aigeneratedproblem/dto/ProblemFile/AiRequest/RequestProblemToAiByTextDto.java
@@ -1,8 +1,7 @@
-package com.app.domain.summary.dto.SummaryFile.AiRequest;
+package com.app.domain.problem.aigeneratedproblem.dto.ProblemFile.AiRequest;
 
 import com.app.global.config.ENUM.Amount;
 import com.app.global.config.ENUM.ProblemDifficulty;
-import com.app.global.config.ENUM.ProblemType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,11 +11,8 @@ import lombok.Setter;
 @NoArgsConstructor
 @Setter
 @Getter
-public class RequestSummaryToAiByTextDto {
+public class RequestProblemToAiByTextDto {
     private String text;
-    private ProblemType type;
     private Amount amount;
     private ProblemDifficulty difficulty;
-    private String fileName;
-
 }
