+++ src/main/java/com/app/domain/problem/membersavedproblem/repository/MemberSavedProblemRepository.java
@@ -1,0 +1,7 @@
+package com.app.domain.problem.membersavedproblem.repository;
+
+import com.app.domain.problem.membersavedproblem.entity.MemberSavedProblem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberSavedProblemRepository extends JpaRepository<MemberSavedProblem, Long> {
+}
