+++ src/main/java/com/app/domain/summary/service/SummaryFileService.java
@@ -4,8 +4,7 @@ package com.app.domain.summary.service;
 import com.app.domain.file.dto.Response.FileListResponseDto;
 import com.app.domain.member.entity.Member;
 import com.app.domain.member.service.MemberService;
-import com.app.domain.problem.dto.ProblemFile.AiRequest.AiGenerateProblemFromAiDto;
-import com.app.domain.problem.value.S3FileInformation;
+import com.app.domain.problem.aigeneratedproblem.value.S3FileInformation;
 import com.app.domain.summary.dto.SummaryFile.Request.AiGenerateSummaryDto;
 import com.app.domain.summary.dto.SummaryFile.AiRequest.AiGenerateSummaryFromAiDto;
 import com.app.domain.summary.dto.SummaryFile.Response.AiGenerateSummaryResponseDto;
@@ -21,12 +20,7 @@ import com.app.global.pdf.SummaryPdfMaker;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.pdfbox.pdmodel.PDDocument;
-import org.apache.pdfbox.pdmodel.PDPage;
-import org.apache.pdfbox.pdmodel.PDPageContentStream;
-import org.apache.pdfbox.pdmodel.font.PDType0Font;
-import org.apache.pdfbox.pdmodel.graphics.state.PDExtendedGraphicsState;
 import org.apache.pdfbox.text.PDFTextStripper;
-import org.jboss.jandex.Main;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.FileSystemResource;
@@ -48,7 +42,6 @@ import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
