+++ src/main/java/com/app/global/pdf/ProblemPdfMaker.java
@@ -1,6 +1,6 @@
 package com.app.global.pdf;
 
-import com.app.domain.problem.dto.ProblemFile.AiRequest.AiGenerateProblemFromAiDto;
+import com.app.domain.problem.aigeneratedproblem.dto.ProblemFile.AiRequest.AiGenerateProblemFromAiDto;
 import com.app.global.config.ENUM.PdfType;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
